+++ lib/paystack_sdk/client.rb
@@ -5,6 +5,7 @@ require_relative "resources/customers"
 require_relative "resources/transfer_recipients"
 require_relative "resources/transfers"
 require_relative "resources/banks"
+require_relative "resources/verification"
 require_relative "utils/connection_utils"
 
 module PaystackSdk
@@ -105,6 +106,20 @@ module PaystackSdk
     # ```
     def banks
       @banks ||= Resources::Banks.new(@connection)
+    end
+
+    # Provides access to the `Verification` resource.
+    #
+    # @return [PaystackSdk::Resources::Verification] An instance of the
+    #  `Verification` resource.
+    #
+    # @example
+    # ```ruby
+    #   verification = client.verification
+    #   response = verification.resolve_account(account_number: ..., bank_code: ...)
+    # ```
+    def verification
+      @verification ||= Resources::Verification.new(@connection)
     end
   end
 end